1) Remove ChannelType that was put on Intent 2) Change to notification helper class 3) showNotification move to helper class
