+++ app/src/main/java/io/github/droidkaigi/confsched2018/presentation/common/broadcastreceiver/NotificationBroadcastReceiver.kt
@@ -4,15 +4,11 @@ import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.support.v4.app.NotificationManagerCompat
 import android.support.v4.app.TaskStackBuilder
-import android.support.v4.content.ContextCompat
-import io.github.droidkaigi.confsched2018.R
 import io.github.droidkaigi.confsched2018.presentation.MainActivity
 import io.github.droidkaigi.confsched2018.presentation.common.pref.Prefs
 import io.github.droidkaigi.confsched2018.presentation.detail.SessionDetailActivity
-import io.github.droidkaigi.confsched2018.util.NotificationUtil.ChannelType
-import io.github.droidkaigi.confsched2018.util.notificationBuilder
+import io.github.droidkaigi.confsched2018.util.NotificationHelper
 import timber.log.Timber
 
 class NotificationBroadcastReceiver : BroadcastReceiver() {
@@ -23,55 +19,38 @@ class NotificationBroadcastReceiver : BroadcastReceiver() {
             return
         }
 
-        Timber.d("Show Notification")
-        val sessionId = intent!!.getStringExtra(EXTRA_SESSION_ID)
-        val title = intent.getStringExtra(EXTRA_TITLE)
-        val text = intent.getStringExtra(EXTRA_TEXT)
-        val channelType = intent.getSerializableExtra(EXTRA_CHANNEL_TYPE) as ChannelType
-        val pendingIntent = TaskStackBuilder
-                .create(context!!)
-                .addNextIntent(MainActivity.createIntent(context))
-                .addNextIntent(SessionDetailActivity.createIntent(context, sessionId))
-                .getPendingIntent(sessionId.hashCode(), PendingIntent.FLAG_UPDATE_CURRENT)
-        showNotification(context, title, text, pendingIntent, channelType)
-    }
-
-    private fun showNotification(
-            context: Context,
-            title: String,
-            text: String,
-            pendingIntent: PendingIntent?,
-            channelId: ChannelType
-    ) {
-        val notification = notificationBuilder(context, channelId)
-                .setContentTitle(title)
-                .setContentText(text)
-                .setFullScreenIntent(pendingIntent, true)
-                .setAutoCancel(true)
-                .setColor(ContextCompat.getColor(context, R.color.primary))
-                .setSmallIcon(R.drawable.ic_notification).build()
-
-        val notificationManagerCompat = NotificationManagerCompat.from(context)
-        notificationManagerCompat.notify(title.hashCode(), notification)
+        when (intent?.action) {
+            ACTION_FAVORITE_SESSION_START -> {
+                Timber.d("Show Notification")
+                val sessionId = intent.getStringExtra(EXTRA_SESSION_ID)
+                val title = intent.getStringExtra(EXTRA_TITLE)
+                val text = intent.getStringExtra(EXTRA_TEXT)
+                val pendingIntent = TaskStackBuilder
+                        .create(context!!)
+                        .addNextIntent(MainActivity.createIntent(context))
+                        .addNextIntent(SessionDetailActivity.createIntent(context, sessionId))
+                        .getPendingIntent(sessionId.hashCode(), PendingIntent.FLAG_UPDATE_CURRENT)
+                NotificationHelper.showFavoriteSessionStart(context, title, text, pendingIntent)
+            }
+        }
     }
 
     companion object {
+        private const val ACTION_FAVORITE_SESSION_START = "ACTION_FAVORITE_SESSION_START"
         private const val EXTRA_SESSION_ID = "EXTRA_SESSION_ID"
         private const val EXTRA_TITLE = "EXTRA_TITLE"
         private const val EXTRA_TEXT = "EXTRA_TEXT"
-        private const val EXTRA_CHANNEL_TYPE = "EXTRA_CHANNEL_TYPE"
         fun createIntent(
                 context: Context,
                 sessionId: String,
                 title: String,
-                text: String,
-                channelType: ChannelType
+                text: String
         ): Intent {
             return Intent(context, NotificationBroadcastReceiver::class.java).apply {
+                action = ACTION_FAVORITE_SESSION_START
                 putExtra(EXTRA_SESSION_ID, sessionId)
                 putExtra(EXTRA_TITLE, title)
                 putExtra(EXTRA_TEXT, text)
-                putExtra(EXTRA_CHANNEL_TYPE, channelType)
             }
         }
     }
